completed the restructuring of files
